# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/permissions": {
            "post": {
                "summary": "创建权限",
                "operationId": "createPermission",
                "requestBody": {
                    "description": "权限数据",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "权限名称",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "权限描述",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/permissions/{id}": {
            "put": {
                "summary": "更新权限",
                "operationId": "updatePermission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "权限ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "权限数据",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "权限名称",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "权限描述",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "删除权限",
                "operationId": "deletePermission",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "权限ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功"
                    }
                }
            }
        },
...
        "/roles/{id}/permissions": {
            "get": {
                "summary": "获取角色权限",
                "operationId": "getRolePermissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            },
            "post": {
                "summary": "分配角色权限",
                "operationId": "assignRolePermissions",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "角色ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "权限ID数组",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission_ids"
                                ],
                                "properties": {
                                    "permission_ids": {
                                        "description": "权限ID数组",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "分配成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "分配成功"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}